+++ Userland/Libraries/LibC/arpa/inet.h
@@ -6,7 +6,6 @@
 
 #pragma once
 
-#include <endian.h>
 #include <inttypes.h>
 #include <netinet/in.h>
 #include <sys/cdefs.h>
@@ -26,33 +25,5 @@ static inline int inet_aton(const char* cp, struct in_addr* inp)
 }
 
 char* inet_ntoa(struct in_addr);
-
-static inline uint16_t htons(uint16_t value)
-{
-#if __BYTE_ORDER == __LITTLE_ENDIAN
-    return __builtin_bswap16(value);
-#else
-    return value;
-#endif
-}
-
-static inline uint16_t ntohs(uint16_t value)
-{
-    return htons(value);
-}
-
-static inline uint32_t htonl(uint32_t value)
-{
-#if __BYTE_ORDER == __LITTLE_ENDIAN
-    return __builtin_bswap32(value);
-#else
-    return value;
-#endif
-}
-
-static inline uint32_t ntohl(uint32_t value)
-{
-    return htonl(value);
-}
 
 __END_DECLS
